Modify authorisation, modify user roles, make modification to task CRUD

+++ src/Controller/TaskController.php
@@ -4,16 +4,19 @@ namespace App\Controller;
 
 use App\Entity\Task;
 use App\Form\TaskType;
+use Doctrine\Persistence\ManagerRegistry;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 
 class TaskController extends AbstractController
 {
     /**
      * @Route("/tasks", name="task_list")
      */
-    public function listAction()
+    public function listAction(): Response
     {
         return $this->render('task/list.html.twig', ['tasks' => $this->getDoctrine()->getRepository('App:Task')->findAll()]);
     }
@@ -21,16 +24,22 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/create", name="task_create")
      */
-    public function createAction(Request $request)
+    public function createAction(ManagerRegistry $managerRegistry, Request $request): RedirectResponse|Response
     {
+
+        $user = $this->getUser();
+
         $task = new Task();
-        $form = $this->createForm(TaskType::class, $task);
+        $form = $this->createForm(TaskType::class, $task, [
+            'validation_groups' => ['create']
+        ]);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
+            $em = $managerRegistry->getManager();
 
+            $task->setUser($user);
             $em->persist($task);
             $em->flush();
 
@@ -45,9 +54,11 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/{id}/edit", name="task_edit")
      */
-    public function editAction(Task $task, Request $request)
+    public function editAction(Task $task, Request $request): RedirectResponse|Response
     {
-        $form = $this->createForm(TaskType::class, $task);
+        $form = $this->createForm(TaskType::class, $task, [
+            'validation_groups' => ['edit']
+        ]);
 
         $form->handleRequest($request);
 
@@ -68,7 +79,7 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/{id}/toggle", name="task_toggle")
      */
-    public function toggleTaskAction(Task $task)
+    public function toggleTaskAction(Task $task): RedirectResponse
     {
         $task->toggle(!$task->isDone());
         $this->getDoctrine()->getManager()->flush();
@@ -81,8 +92,22 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/{id}/delete", name="task_delete")
      */
-    public function deleteTaskAction(Task $task)
+    public function deleteTaskAction(Task $task): RedirectResponse
     {
+        $user = $this->getUser();
+
+        $taskUser = $task->getUser();
+
+        if (!is_null($user)) {
+            $userRoles = $user->getRoles();
+        }
+
+        if (is_null($user) || ($taskUser !== $user && (!in_array('ROLE_ADMIN',$userRoles) && $taskUser->getEmail() === 'anon@test.com'))) {
+            $this->addFlash('error', sprintf('La tâche %s peut être supprimée uniquement par l\'utilisateur qui l\'a créé.', $task->getTitle()));
+
+            return $this->redirectToRoute('task_list');
+        }
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($task);
         $em->flush();
